VZ-1306: Break up testutils.go
