+++ pkg/managed/istio_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/verrazzano/verrazzano-operator/pkg/constants"
-	"github.com/verrazzano/verrazzano-operator/pkg/testutils"
+	"github.com/verrazzano/verrazzano-operator/pkg/testutil"
 	"github.com/verrazzano/verrazzano-operator/pkg/util"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -24,7 +24,7 @@ import (
 )
 
 func TestGetUniqueServiceEntryAddress(t *testing.T) {
-	clusterConnections := testutils.GetManagedClusterConnections()
+	clusterConnections := testutil.GetManagedClusterConnections()
 	clusterConnection := clusterConnections["cluster1"]
 	var startIPIndex = 1
 
@@ -60,8 +60,8 @@ func TestGetUniqueServiceEntryAddress(t *testing.T) {
 }
 
 func TestGetIstioGateways(t *testing.T) {
-	modelBindingPair := testutils.GetModelBindingPair()
-	clusterConnections := testutils.GetManagedClusterConnections()
+	modelBindingPair := testutil.GetModelBindingPair()
+	clusterConnections := testutil.GetManagedClusterConnections()
 	clusterConnection := clusterConnections["cluster2"]
 
 	gatewayAddress := getIstioGateways(modelBindingPair, clusterConnections, "cluster2")
@@ -90,8 +90,8 @@ func TestGetIstioGateways(t *testing.T) {
 }
 
 func TestCleanupOrphanedServiceEntries(t *testing.T) {
-	modelBindingPair := testutils.GetModelBindingPair()
-	clusterConnections := testutils.GetManagedClusterConnections()
+	modelBindingPair := testutil.GetModelBindingPair()
+	clusterConnections := testutil.GetManagedClusterConnections()
 	clusterConnection := clusterConnections["cluster1"]
 	clusterConnection2 := clusterConnections["cluster3"]
 
@@ -127,8 +127,8 @@ func TestCleanupOrphanedServiceEntries(t *testing.T) {
 }
 
 func TestCleanupOrphanedIngresses(t *testing.T) {
-	modelBindingPair := testutils.GetModelBindingPair()
-	clusterConnections := testutils.GetManagedClusterConnections()
+	modelBindingPair := testutil.GetModelBindingPair()
+	clusterConnections := testutil.GetManagedClusterConnections()
 	clusterConnection := clusterConnections["cluster1"]
 	clusterConnection2 := clusterConnections["cluster3"]
 
@@ -179,8 +179,8 @@ func TestCleanupOrphanedIngresses(t *testing.T) {
 }
 
 func TestCreateServiceEntries(t *testing.T) {
-	modelBindingPair := testutils.GetModelBindingPair()
-	clusterConnections := testutils.GetManagedClusterConnections()
+	modelBindingPair := testutil.GetModelBindingPair()
+	clusterConnections := testutil.GetManagedClusterConnections()
 	clusterConnection := clusterConnections["cluster1"]
 
 	// create the service entries
@@ -257,8 +257,8 @@ func assertCreateServiceEntries(t *testing.T, clusterConnection *util.ManagedClu
 }
 
 func TestCreateIngresses(t *testing.T) {
-	modelBindingPair := testutils.GetModelBindingPair()
-	clusterConnections := testutils.GetManagedClusterConnections()
+	modelBindingPair := testutil.GetModelBindingPair()
+	clusterConnections := testutil.GetManagedClusterConnections()
 	clusterConnection := clusterConnections["cluster1"]
 
 	// create the ingresses
@@ -311,8 +311,8 @@ func assertCreateIngresses(t *testing.T, clusterConnection *util.ManagedClusterC
 }
 
 func TestCreateDestinationRules(t *testing.T) {
-	modelBindingPair := testutils.GetModelBindingPair()
-	clusterConnections := testutils.GetManagedClusterConnections()
+	modelBindingPair := testutil.GetModelBindingPair()
+	clusterConnections := testutil.GetManagedClusterConnections()
 	clusterConnection := clusterConnections["cluster1"]
 
 	// create the destination rules
@@ -379,8 +379,8 @@ func assertCreateDestinationRules(t *testing.T, clusterConnection *util.ManagedC
 }
 
 func TestCreateAuthorizationPolicies(t *testing.T) {
-	clusterConnections := testutils.GetManagedClusterConnections()
-	modelBindingPair := testutils.GetModelBindingPair()
+	clusterConnections := testutil.GetManagedClusterConnections()
+	modelBindingPair := testutil.GetModelBindingPair()
 	clusterConnection := clusterConnections["cluster1"]
 
 	// create the authorization policies
@@ -417,8 +417,8 @@ func assertCreateAuthorizationPolicies(t *testing.T, clusterConnection *util.Man
 	assert.Equal(t, 1, len(policy.Spec.Rules))
 	assert.Equal(t, 2, len(policy.Spec.Rules[0].From))
 	assert.Equal(t, 2, len(policy.Spec.Rules[0].From[0].Source.Namespaces))
-	assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test"))
-	assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
+	assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test"))
+	assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
 
 	// test2-authorization-policy
 	list, err = clusterConnection.IstioAuthClientSet.SecurityV1beta1().AuthorizationPolicies("test2").List(context.TODO(), metav1.ListOptions{})
@@ -435,13 +435,13 @@ func assertCreateAuthorizationPolicies(t *testing.T, clusterConnection *util.Man
 	assert.Equal(t, 2, len(policy.Spec.Rules[0].From))
 	if wlsHasConn {
 		assert.Equal(t, 3, len(policy.Spec.Rules[0].From[0].Source.Namespaces))
-		assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test2"))
-		assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test3"))
-		assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
+		assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test2"))
+		assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test3"))
+		assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
 	} else {
 		assert.Equal(t, 2, len(policy.Spec.Rules[0].From[0].Source.Namespaces))
-		assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test2"))
-		assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
+		assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test2"))
+		assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
 	}
 
 	// test3-authorization-policy
@@ -458,9 +458,9 @@ func assertCreateAuthorizationPolicies(t *testing.T, clusterConnection *util.Man
 	assert.Equal(t, 1, len(policy.Spec.Rules))
 	assert.Equal(t, 2, len(policy.Spec.Rules[0].From))
 	assert.Equal(t, 3, len(policy.Spec.Rules[0].From[0].Source.Namespaces))
-	assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test"))
-	assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test3"))
-	assert.True(t, testutils.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
+	assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test"))
+	assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "test3"))
+	assert.True(t, testutil.StringArrayContains(policy.Spec.Rules[0].From[0].Source.Namespaces, "istio-system"))
 }
 
 func TestNewIngresses(t *testing.T) {
